+++ models/__init__.py
@@ -4,6 +4,7 @@ from . import sendpulse_message
 from . import sendpulse_facebook_page
 from . import sendpulse_faq_entry
 from . import sendpulse_public_template
+from . import sendpulse_privacy_consent_log
 from . import sendpulse_connect
 from . import sendpulse_identify_wizard
 from . import mail_channel